pyproject.toml
[project]
name = "solid-design-patterns"
version = "0.1.0"
description = "Curso de Principios SOLID y Patrones de Diseño en Python"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "stripe>=10.12.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "ruff>=0.6.8",
]


[tool.ruff]
line-length = 79